refactor errorHandler package to errors
