+++ errors/errors.go
@@ -1,4 +1,4 @@
-package errorHandler
+package errors
 
 import (
 	"fmt"
@@ -54,7 +54,7 @@ func BuildErrorMessage(err error) error {
 		// retryable error codes - [502, 503, 504]
 		errString = fmt.Sprintf("The server returned a %s error (%v). Please wait a few minutes and try again.", http.StatusText(errCode), errCode)
 	} else {
-		// non-retryable errorHandler
+		// non-retryable errors
 		errString = fmt.Sprintf("The server returned a %s error (%v). Please contact Turbot support.", http.StatusText(errCode), errCode)
 	}
 	return errors.New(errString)
