+++ turbot/resource_turbot_folder_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/terraform-providers/terraform-provider-turbot/apiClient"
-	"github.com/terraform-providers/terraform-provider-turbot/errorHandler"
+	"github.com/terraform-providers/terraform-provider-turbot/errors"
 	"testing"
 )
 
@@ -226,7 +226,7 @@ func testAccCheckFolderDestroy(s *terraform.State) error {
 			if err == nil {
 				return fmt.Errorf("alert still exists")
 			}
-			if !errorHandler.NotFoundError(err) {
+			if !errors.NotFoundError(err) {
 				return fmt.Errorf("expected 'not found' error, got %s", err)
 			}
 		}
