+++ turbot/resource_turbot_grant.go
@@ -3,7 +3,7 @@ package turbot
 import (
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/terraform-providers/terraform-provider-turbot/apiClient"
-	"github.com/terraform-providers/terraform-provider-turbot/errorHandler"
+	"github.com/terraform-providers/terraform-provider-turbot/errors"
 )
 
 // map of Terraform properties to Turbot properties that we pass to create and update mutations
@@ -130,7 +130,7 @@ func resourceTurbotGrantRead(d *schema.ResourceData, meta interface{}) error {
 
 	Grant, err := client.ReadGrant(id)
 	if err != nil {
-		if errorHandler.NotFoundError(err) {
+		if errors.NotFoundError(err) {
 			// Grant was not found - clear id
 			d.SetId("")
 		}
