+++ turbot/resource_turbot_grant_activation_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/terraform-providers/terraform-provider-turbot/apiClient"
-	"github.com/terraform-providers/terraform-provider-turbot/errorHandler"
+	"github.com/terraform-providers/terraform-provider-turbot/errors"
 	"testing"
 )
 
@@ -96,7 +96,7 @@ func testAccCheckLocalGrantDestroy(s *terraform.State) error {
 			if err == nil {
 				return fmt.Errorf("Alert still exists")
 			}
-			if !errorHandler.NotFoundError(err) {
+			if !errors.NotFoundError(err) {
 				return fmt.Errorf("expected 'not found' error, got %s", err)
 			}
 		}
@@ -133,7 +133,7 @@ func testAccCheckActiveGrantDestroy(s *terraform.State) error {
 		if err == nil {
 			return fmt.Errorf("Alert still exists")
 		}
-		if !errorHandler.NotFoundError(err) {
+		if !errors.NotFoundError(err) {
 			return fmt.Errorf("expected 'not found' error, got %s", err)
 		}
 	}
