+++ turbot/resource_turbot_group_profile.go
@@ -3,7 +3,7 @@ package turbot
 import (
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/terraform-providers/terraform-provider-turbot/apiClient"
-	"github.com/terraform-providers/terraform-provider-turbot/errorHandler"
+	"github.com/terraform-providers/terraform-provider-turbot/errors"
 	"github.com/terraform-providers/terraform-provider-turbot/helpers"
 )
 
@@ -79,7 +79,7 @@ func resourceTurbotGroupProfileRead(d *schema.ResourceData, meta interface{}) er
 
 	groupProfile, err := client.ReadGroupProfile(id)
 	if err != nil {
-		if errorHandler.NotFoundError(err) {
+		if errors.NotFoundError(err) {
 			// profile was not found - clear id
 			d.SetId("")
 		}
