+++ turbot/resource_turbot_resource.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/terraform-providers/terraform-provider-turbot/apiClient"
-	"github.com/terraform-providers/terraform-provider-turbot/errorHandler"
+	"github.com/terraform-providers/terraform-provider-turbot/errors"
 	"github.com/terraform-providers/terraform-provider-turbot/helpers"
 	"strings"
 )
@@ -137,7 +137,7 @@ func resourceTurbotResourceRead(d *schema.ResourceData, meta interface{}) error 
 	// read full resource
 	resource, err := client.ReadFullResource(id)
 	if err != nil {
-		if errorHandler.NotFoundError(err) {
+		if errors.NotFoundError(err) {
 			// resource was not found - clear id
 			d.SetId("")
 		}
